[328801] Fix filled ellipse rendering
When rendering a filled ellipse on a scaled canvas, a small gap
can appear between the fill and the outline shape.
This happens because the outline of the Ellipse is drawn using the
scaled rectangle returned by zoomRect while the inner circle is drawn
using the rectangle returned by zoomFillrect and those rectangle centers
are not aligned.
This patch align those centers to fix this problem.

Signed-off-by: Jules Clero <julesx.clero@intel.com>

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=328801
Also-by: Sebastien Delestaing <sebastien.delestaing@intel.com>